+++ play-frontend-hmrc-play-30/src/main/scala/uk/gov/hmrc/govukfrontend/views/viewmodels/input/Input.scala
@@ -18,11 +18,32 @@ package uk.gov.hmrc.govukfrontend.views.viewmodels.input
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
-import uk.gov.hmrc.govukfrontend.views.viewmodels.CommonJsonFormats._
+import uk.gov.hmrc.govukfrontend.views.viewmodels.FormGroup
 import uk.gov.hmrc.govukfrontend.views.viewmodels.errormessage.ErrorMessage
 import uk.gov.hmrc.govukfrontend.views.viewmodels.hint.Hint
 import uk.gov.hmrc.govukfrontend.views.viewmodels.label.Label
 
+/** Parameters to `GovukInput` Twirl template
+  *
+  * @param id `id` attribute for the `input`
+  * @param name `name` attribute for the `input`
+  * @param inputType `type` attribute for the `input`
+  * @param inputmode optional `inputmode` attribute for the `input`
+  * @param describedBy optional `aria-describedby` attribute for the `input`
+  * @param value optional `value` attribute for the `input`
+  * @param label optional `Label` for the control
+  * @param hint optional `Hint` for the control
+  * @param errorMessage optional `ErrorMessage` to display
+  * @param formGroup additional CSS classes/attributes/etc. to apply to the form group
+  * @param classes optional additional CSS classes to apply to the `input`
+  * @param autocomplete optional `autocomplete` attribute for the `input`
+  * @param pattern optional `pattern` attribute for the `input`
+  * @param attributes optional additional HTML attributes to apply to the `input`
+  * @param spellcheck optional `spellcheck` attribute for the `input`
+  * @param prefix optional content to display immediately before the `input`
+  * @param suffix optional content to display immediately after the `input`
+  * @param disabled optional `disabled` attribute for the `input`
+  */
 case class Input(
   id: String = "",
   name: String = "",
@@ -33,7 +54,7 @@ case class Input(
   label: Label = Label(),
   hint: Option[Hint] = None,
   errorMessage: Option[ErrorMessage] = None,
-  formGroupClasses: String = "",
+  formGroup: FormGroup = FormGroup.empty,
   classes: String = "",
   autocomplete: Option[String] = None,
   pattern: Option[String] = None,
@@ -59,7 +80,7 @@ object Input {
         (__ \ "label").readWithDefault[Label](defaultObject.label) and
         (__ \ "hint").readNullable[Hint] and
         (__ \ "errorMessage").readNullable[ErrorMessage] and
-        readsFormGroupClasses and
+        (__ \ "formGroup").readWithDefault[FormGroup](defaultObject.formGroup) and
         (__ \ "classes").readWithDefault[String](defaultObject.classes) and
         (__ \ "autocomplete").readNullable[String] and
         (__ \ "pattern").readNullable[String] and
@@ -81,7 +102,7 @@ object Input {
         (__ \ "label").write[Label] and
         (__ \ "hint").writeNullable[Hint] and
         (__ \ "errorMessage").writeNullable[ErrorMessage] and
-        writesFormGroupClasses and
+        (__ \ "formGroup").write[FormGroup] and
         (__ \ "classes").write[String] and
         (__ \ "autocomplete").writeNullable[String] and
         (__ \ "pattern").writeNullable[String] and