+++ play-frontend-hmrc-play-30/src/test/scala/uk/gov/hmrc/govukfrontend/views/viewmodels/Generators.scala
@@ -22,6 +22,16 @@ import play.twirl.api.{Html, HtmlFormat}
 
 object Generators {
 
+  val arbFormGroup: Arbitrary[FormGroup] = Arbitrary {
+    for {
+      classes    <- Gen.option(genClasses())
+      attributes <- genAttributes()
+    } yield FormGroup(
+      classes = classes,
+      attributes = attributes
+    )
+  }
+
   val genNonEmptyAlphaStr: Gen[String] = Gen.alphaStr.suchThat(_.nonEmpty)
 
   /**
@@ -37,14 +47,14 @@ object Generators {
     value <- genNonEmptyAlphaStr
   } yield (attr, value)
 
-  def genAttributes(nAttributes: Int = 5) =
+  def genAttributes(nAttributes: Int = 5): Gen[Map[String, String]] =
     for {
       sz         <- Gen.chooseNum(0, nAttributes)
       attributes <- Gen.mapOfN[String, String](sz, genAttrVal)
     } yield attributes
 
   val genHtmlString: Gen[String] =
-    Gen.oneOf(Gen.const("""<p>some paragraph</p>"""), Gen.const(""""<b>Back</b>""""))
+    Gen.oneOf(Gen.const("""<p>some paragraph</p>"""), Gen.const("""<b>Back</b>"""))
 
   implicit val arbHtml: Arbitrary[Html] = Arbitrary {
     for {