+++ play-frontend-hmrc-play-30/src/test/scala/uk/gov/hmrc/govukfrontend/views/viewmodels/charactercount/Generators.scala
@@ -26,21 +26,21 @@ object Generators {
 
   implicit val arbCharacterCount: Arbitrary[CharacterCount] = Arbitrary {
     for {
-      id               <- genNonEmptyAlphaStr
-      name             <- genNonEmptyAlphaStr
-      rows             <- Gen.chooseNum(0, 5)
-      value            <- Gen.option(genAlphaStr())
-      maxLength        <- Gen.option(Gen.chooseNum(1, 10))
-      maxWords         <- Gen
-                            .option(Gen.chooseNum(1, 10))
-                            .retryUntil(optMaxWords => maxLength.nonEmpty || optMaxWords.nonEmpty)
-      threshold        <- Gen.option(Gen.chooseNum(1, 5))
-      label            <- arbLabel.arbitrary
-      hint             <- Gen.option(arbHint.arbitrary)
-      errorMessage     <- Gen.option(arbErrorMessage.arbitrary)
-      formGroupClasses <- genClasses()
-      classes          <- genClasses()
-      attributes       <- genAttributes()
+      id           <- genNonEmptyAlphaStr
+      name         <- genNonEmptyAlphaStr
+      rows         <- Gen.chooseNum(0, 5)
+      value        <- Gen.option(genAlphaStr())
+      maxLength    <- Gen.option(Gen.chooseNum(1, 10))
+      maxWords     <- Gen
+                        .option(Gen.chooseNum(1, 10))
+                        .retryUntil(optMaxWords => maxLength.nonEmpty || optMaxWords.nonEmpty)
+      threshold    <- Gen.option(Gen.chooseNum(1, 5))
+      label        <- arbLabel.arbitrary
+      hint         <- Gen.option(arbHint.arbitrary)
+      errorMessage <- Gen.option(arbErrorMessage.arbitrary)
+      formGroup    <- arbFormGroup.arbitrary
+      classes      <- genClasses()
+      attributes   <- genAttributes()
     } yield CharacterCount(
       id = id,
       name = name,
@@ -52,7 +52,7 @@ object Generators {
       label = label,
       hint = hint,
       errorMessage = errorMessage,
-      formGroupClasses = formGroupClasses,
+      formGroup = formGroup,
       classes = classes,
       attributes = attributes
     )